bridge_track: set MSTI port STP states
Now that we have assigned VLANs to MSTIs, we can set MSTI STP states. We
can only set states for MSTIs for which at least one VLAN is assigned to
a port though, so we need to check for that first.

Likewise, if a port gets assigned a new VLAN, we need to check if we
need to update the kernel MSTI STP state, and if a VLAN gets moved to a
new MSTI, we may need to set the STP state for ports as well.

Co-authored-by: Tomas Kyzlink <tkyzlink@dtitech.cz>
Signed-off-by: Tomáš Kyzlink <tkyzlink@dtitech.cz>